perf(hyperdim): implement branchless bitmask packing and refactor tests
- Replace branchy thresholding with branchless bitmask construction in
  `BundleAccumulator::finalize` and `HVec10240::bundle`.
- Move `hyperdim` unit tests to `src/hyperdim/hyperdim_tests.rs` to
  comply with 500 LOC limit.

Measured speedups:
- Finalize: ~40%
- Bundling: ~14-20%
